Rename the whole Views/* area to CentralUI/*
This reflects the versatile nature of that subdirectory far better (It
wasn't only about Views quiete for a while now)
