+++ src/libHeaven/App/ModeSwitcher.hpp
@@ -5,7 +5,7 @@
 #include <QHash>
 #include <QStringList>
 
-#include "Views/WindowState.hpp"
+#include "CentralUI/WindowState.hpp"
 
 namespace Heaven
 {
