+++ src/libHeaven/CentralUI/Contexts/ContextKeys.cpp
@@ -17,7 +17,7 @@
 #include <QMap>
 #include <QStringBuilder>
 
-#include "Views/Contexts/ContextKeys.hpp"
+#include "CentralUI/Contexts/ContextKeys.hpp"
 
 namespace Heaven
 {
