+++ src/libHeaven/CentralUI/Contexts/ViewContext.cpp
@@ -1,6 +1,8 @@
 /*
  * libHeaven - A Qt-based ui framework for strongly modularized applications
- * Copyright (C) 2012-2013 Sascha Cunz <sascha@babbelbox.org>
+ * Copyright (C) 2012-2013 The MacGitver-Developers <dev@macgitver.org>
+ *
+ * (C) Sascha Cunz <sascha@macgitver.org>
  *
  * This program is free software; you can redistribute it and/or modify it under the terms of the
  * GNU General Public License (Version 2) as published by the Free Software Foundation.
@@ -14,13 +16,30 @@
  *
  */
 
-#include "libHeaven/Views/GlobalView.hpp"
+#include "libHeaven/CentralUI/Contexts/ViewContext.hpp"
 
 namespace Heaven
 {
 
-    GlobalView::GlobalView( const QString& identifier )
-        : ContextView( identifier, GlobalViewType )
+    ViewContext::ViewContext( ContextView* owningView )
+        : mOwningView( owningView )
+    {
+    }
+
+    ViewContext::~ViewContext()
+    {
+    }
+
+    ContextView* ViewContext::owningView()
+    {
+        return mOwningView;
+    }
+
+    void ViewContext::afterDetached()
+    {
+    }
+
+    void ViewContext::beforeAttach()
     {
     }
