+++ src/libHeaven/CentralUI/Views/View.cpp
@@ -17,7 +17,7 @@
 #include <QToolBar>
 #include <QVBoxLayout>
 
-#include "libHeaven/Views/View.h"
+#include "libHeaven/CentralUI/Views/View.hpp"
 
 namespace Heaven
 {
