--- src/libHeaven/Views/DependantView.cpp
@@ -1,50 +1,0 @@
-/*
- * libHeaven - A Qt-based ui framework for strongly modularized applications
- * Copyright (C) 2012-2013 The MacGitver-Developers <dev@macgitver.org>
- *
- * (C) Sascha Cunz <sascha@macgitver.org>
- *
- * This program is free software; you can redistribute it and/or modify it under the terms of the
- * GNU General Public License (Version 2) as published by the Free Software Foundation.
- *
- * This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
- * even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with this program; if
- * not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-#include "libHeaven/Views/DependantView.hpp"
-#include "libHeaven/Views/Contexts/ViewContextManager.hpp"
-
-namespace Heaven
-{
-
-    DependantView::DependantView( const QString& identifier )
-        : ContextView( identifier, DependantViewType )
-    {
-    }
-
-    void DependantView::setDependency( const QString& identifier )
-    {
-        if( !mDependency.isEmpty() )
-        {
-            ViewContextManager::self().unregisterDependency( this, mDependency );
-        }
-
-        mDependency = identifier;
-
-        if( !mDependency.isEmpty() )
-        {
-            ViewContextManager::self().registerDependency( this, mDependency );
-        }
-    }
-
-    QString DependantView::dependency() const
-    {
-        return mDependency;
-    }
-
-}
